refactor: Moved alert component to a new file
